+++ packages/terra-infinite-list/src/terra-dev-site/doc/example/mock-data/virtualizedData.js
@@ -1,4 +1,3 @@
-
 const mockData = (() => {
   const items = [];
   for (let i = 0; i < 100; i += 1) {