+++ packages/terra-brand-footer/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated `master` references to `main`.
 
 2.24.0 - (May 26, 2020)
 ------------------