+++ packages/terra-brand-footer/CHANGELOG.md
@@ -7,6 +7,9 @@ Unreleased
 * `links` prop has been deprecated, will resolve to `sections`.
 * Updated UpgradeGuide.
 
+### Changed
+* updated `master` references to `main`.
+
 2.24.0 - (May 26, 2020)
 ------------------
 ### Changed