+++ packages/terra-brand-footer/CHANGELOG.md
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ----------
+### Breaking Changes
+* `links` prop has been deprecated, will resolve to `sections`.
+* Updated UpgradeGuide.
+
 ### Changed
 * updated `master` references to `main`.