+++ packages/terra-brand-footer/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+3.0.0 - (July 14, 2020)
+------------------
 ### Breaking Changes
 * `links` prop has been deprecated, will resolve to `sections`.
 * Updated UpgradeGuide.